% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arg4_res}
\alias{arg4_res}
\title{Argument 4: simulations}
\format{
A list containing the mean for both scenarios,
}
\usage{
arg4_res
}
\description{
This data set is the one used to draw Figure 4. As these simulation take
~1hour to run (1CPU - Intel i7), we saved the results to speed up the
creation of Figure 4 but note that \code{\link[=scr_arg4]{scr_arg4()}} allows the reproduction of
those results.
}
\keyword{datasets}
