% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coOccProb.R
\name{coOccProb}
\alias{coOccProb}
\title{Calculate probability of each event not happening}
\usage{
coOccProb(pA, pB, covAB)
}
\arguments{
\item{pA}{probability of event A.}

\item{pB}{probability of event B.}

\item{covAB}{covariance between A and B.}
}
\description{
Calculate probability of each event not happening
}
\examples{
coOccProb(.5, .5, 0) 
}
