% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_arg1.R
\name{scr_arg1}
\alias{scr_arg1}
\title{Argument 1: environment and occurrence}
\usage{
scr_arg1(denv = 0.01)
}
\arguments{
\item{denv}{incremental value for the environmental gradient.}
}
\description{
Code to reproduce analysis for 1st argument and figure 1.
}
