% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_arg2.R
\name{scr_arg2}
\alias{scr_arg2}
\title{Argument 2: co-occurrence with 3 species}
\usage{
scr_arg2(nval = 100)
}
\arguments{
\item{nval}{Number of values used as x values.}
}
\description{
Code to reproduce analysis for argument 2 and figure 2.
}
