% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_arg4.R
\name{scr_arg4}
\alias{scr_arg4}
\title{Argument 4: co-occurrence and scale}
\usage{
scr_arg4(run_all = FALSE, nrep = 10000)
}
\arguments{
\item{run_all}{Number of repetitions for every environmental values.}

\item{nrep}{Number of repetitions for every environmental values, ignored if
'\code{run_all = FALSE}.}
}
\description{
Code to reproduce analysis for argument 4 and figure 4. Use
\code{set.seed(7891)} and \code{nrep = 1e-5} for to obtain the results` found in the
original study.
}
