% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_arg6.R
\name{scr_arg6}
\alias{scr_arg6}
\title{Argument 6: co-occurrence and co-existence}
\usage{
scr_arg6(nval = 200)
}
\arguments{
\item{nval}{Number of interactions strength values.}
}
\description{
Code to reproduce analysis for argument 6 and figure 6.
}
