% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_arg7.R
\name{scr_arg7}
\alias{scr_arg7}
\title{Argument 7: co-occurrence and co-existence}
\usage{
scr_arg7(nval = 200)
}
\arguments{
\item{nval}{Number of interactions strength values.}
}
\description{
Code to reproduce analysis for argument 7 and figure 6.
}
